Fix lsp-zero config
A couple of language servers need to be manually installed in order to
use the correct version:
- ruby_lsp
- gopls
- rust_analyzer

This changes the lsp-config to not auto-install and initialize them.

Additionally, I'm using the dev version of nvim (0.10.x), so the extra
config for ruby is no longer necessary, and was actually breaking
semantic highlighting.