Simplify .gitignore
Previously, I was manually adding every dir to be ignored into the
.gitignore file. This is both tedious, and unrealistic to maintain
long-term. Instead, it is now ignoring everything in the root dir,
and wanted dirs must be whitelisted to be checked into git. This is
much simpler to maintain, especially considering only one dir is even
checked into git at this point.