Workaround TS code actions
The current config for TypeScript code actions in LSP aren't working
quite right, and causing there to be unsaved changes again after every
time I save. For now, this removes the extra code actions on the buffer
write so that it doesn't do that, but I still need to get to the bottom
of it.